Change `skip_download` to `skip` and improve usage
Renamed the `skip_download` param to simply `skip`.
It is only supported from `in`, and will skip fetching
the resource all together.

Also improved the handling of this scenario:

* exits immediately with `0` (instead of if/else switching)
* returns 'skipped' in the `status` metadata
* prints out message to console indicating it was skipped

Again, the primary motivation for this is to support skipping the
implicit `get` that follows every `put`.  Some `put` endpoints may not
be able to respond to `get`, and we don't want to error the pipeline for
that.