New `hash` and `default` version config options
This adds new `hash` and `default` options to the `source.version` config.

```yaml
  source:
    version:
      hash: {headers|body|full}
      default: {hash|none}
```

The precedence is now:

1. `jq`
2. `header`
3. `hash`
4. `default`

If `default: none` is used, and no other version strategy was able to
yield a non-empty string value, then no new version will be returned from
`check`.  This is most useful in scenarios where you want to version the
resource on a response header or `jq` query that may not be present in
every response from the endpoint.  In those scenarios, it may be
desireable to skip that version (rather than default to generating a
`hash` for the version).  For example,

```yaml
  source:
    version:
      jq: .may.not.always.be.present.property
      default: none
```