Add test coverage for data file usage
Added test coverage for `out` operations that use a data file in the
`curl` request (`-d`).

This required stubbing out the `replaceBuildMetadataInFile()` method
due to the difficulty of trying to mock this behavior with a real file.
We would have had to coordinate the test config fixture with the
`$BATS_TEST_TMPDIR` location and that would be more effort
(and confusion) than the test is currently worth.

But these new tests at least give us coverage to assert we use the data
file config if/when it is configured in either the `source` or as a step
`params`.