Make build metadata replacements an "opt-in" feature
Prior to this, the resource would always implicitly perform build
metadata token replacements on all headers and body content (regardless
of whether the body was flagged as `data_binary` or not).

This can create unnecessary trouble for scenarios where you are working
with a large body, or if the body is binary (in which case token
replacement doesn't make sense anyways).

In this commit, we are now making the build metadata token replacement
something you must opt-in for via the `build_metadata` config. It will
no longer attempt to replace the tokens by default, and accepts a list
of any/all of the following:

* `headers` - replace any tokens configured in the `headers`.
* `body` - replace any tokens in either `file` or `text` content.

For example, to `POST` some inline `text` content and headers that use
metadata tokens:

```yaml
resources:
  - name: http-bin-post
    type: http-resource
    source:
      url: https://httpbin.org/post

jobs:
  - name: post-something
    plan:
      - put: http-bin-post
        params:
          build_metadata: [headers,body]
          headers:
            Content-Type: application/json
            X-Build-Logs: $ATC_EXTERNAL_URL/teams/$BUILD_TEAM_NAME/pipelines/$BUILD_PIPELINE_NAME/jobs/$BUILD_JOB_NAME/builds/$BUILD_NAME
  	  text: |
            See build logs at $ATC_EXTERNAL_URL/teams/$BUILD_TEAM_NAME/pipelines/$BUILD_PIPELINE_NAME/jobs/$BUILD_JOB_NAME/builds/$BUILD_NAME.
```

The `build_metadata` can also be declared in the `source` config for the
resource (which would enable it for `check` operations as well), and can
be overridden in each `get` or `put` step.

See the README for more details.

Closes #7