Support HEAD method by implicitly including '--head' option
We use curl under the hood to perform the http invocation.  Curl is
quite particular with the HEAD method (`-X HEAD`) and strongly suggests
you specify the `--head` option.

Attempting without it:

    curl -X HEAD https://www.wikipedia.org
    Warning: Setting custom HTTP method to HEAD with -X/--request may not work the
    Warning: way you want. Consider using -I/--head instead.

This commit adds some special handling that will add the `--head` option
if the method is HEAD (case insensitive).

Closes #8